+++ src/game/declarative_effects.rs
@@ -755,6 +755,7 @@ impl Game {
             })
             | EffectDef::CannotBeForcedToSacrifice
             | EffectDef::ReduceGenericCostBy(_)
+            | EffectDef::IncreaseMatchingAbilityCostBy { .. }
             | EffectDef::IncreaseMatchingSpellCostBy { .. }
             | EffectDef::ReduceMatchingSpellCostBy { .. }
             | EffectDef::LandwalkCanBeBlocked(_)