+++ src/game/tests.rs
@@ -579,6 +579,7 @@ mod extra_turns;
 mod fetch_tapped;
 mod filtered_player_prevention;
 mod ghouls;
+mod gloom;
 mod guardian_beast;
 mod held_tapped;
 mod howling_mine;